backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
psycopg[binary]
celery
redis
gunicorn
requests
xmlrpc
